connect to MongoDB before starting server.
Create and export new function called seedDatabaseExpress from seed.js. This function is speicifcally for express routes, but basically is the original seedDatabase static node.js function.

Can now send data from frontend to backend and upload a listing to the database. (Images field still not implemented, so when uploading images needs to be blank).